pyproject.toml
[project]
name = "line-bot-butler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "line-bot-sdk>=3.16.1",
    "ollama>=0.4.7",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.29",
    "schedule>=1.2.2",
]

[tool.mypy]
ignore_missing_imports = true
